% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(formu = f.species, df = track, analysis = "glm",
  n_bootstrap = 100, pop_size = pop.size, stratified = F,
  bayesianInput_ = NA)
}
\arguments{
\item{formu}{formula to be used in the analysis}

\item{df}{data frame to be used in the analysis}

\item{analysis}{type of the analysis, "glm" and "glmm" currently available, "bayes" coming soon}

\item{n_bootstrap}{number of bootstraps}

\item{pop_size}{number of individuals}

\item{stratified}{logical wether a stratified or a pooled bootstrap should be performed}

\item{bayesianInput_}{a list containing a Bayesian Setup, settings, and a character specifying the sampler created with Bayesian Tools}
}
\value{
returns betas of the bootstrapped analysis
}
\description{
This function generates bootstrapped betas.
}
\examples{
bootstrap()

}
\keyword{bootstrap}
