% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_install.R
\name{load.and.install}
\alias{load.and.install}
\title{A unified loader for github and CRAN packages}
\usage{
load.and.install(required.packages, mode = "CRAN", update = F)
}
\arguments{
\item{required.packages}{a vector of characters}

\item{mode}{character defining wether the packages are provided by CRAN or github, use "CRAN" or "GIT"}

\item{update}{logical wether to force updating previously installed packages}
}
\value{
none
}
\description{
This function installs, updates and loads packages provided by CRAN or github
}
\examples{
load.and.install(c("ggplot2", "raster"), mode = "CRAN", update = T)

}
\keyword{load}
\keyword{packages,}
