% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement_generation_stepwise_dopar.R
\name{move.from.preds.stepwise}
\alias{move.from.preds.stepwise}
\title{A resource selection simulator}
\usage{
move.from.preds.stepwise(formu, pop_size = pop.size, steps_ = steps,
  non_steps = non.steps, base_, n_preds, n_preds_original,
  gridsize = sqrt(nrow(predictors)), predictors_, ind_sd = NA,
  dist_effect = NA, stepwise = F, absence_sampling = NA,
  param_list = list(NA))
}
\arguments{
\item{formu}{formula without an response such as ~ x1 + x2}

\item{pop_size}{integer defining the number of individuals}

\item{steps_}{number of relocations/steps}

\item{non_steps}{number of absences}

\item{base_}{vector of numerics as the parameters coefficients}

\item{n_preds}{integer defining total number of predictors, quadratic conditions and interactions}

\item{n_preds_original}{integer defining the number of predictors}

\item{gridsize}{integer determining size of the map}

\item{predictors_}{matrix of predictors}

\item{ind_sd}{integer as the standard deviation of the normal distribution with a mean of 1 that is multiplied with the parameter coefficients}

\item{dist_effect}{numeric defining the impact of the distance on relocation as plogis(1/dist^dist_effect)}

\item{stepwise}{logical, if relocations are only dependent of the set of predictors  or also dependent of the location of the last step}

\item{absence_sampling}{default is "availability" , meaning all cells are used as availability background. "absence" samples only cells where no presence has been recorded.}

\item{pred_choice}{vector  of integers defining the set of predictors originally selected}
}
\value{
returns a dataset usable in resource selection functions and a map suitable for plotting with ggplot2
}
\description{
This function generates a dataset usable in resource selection functions
}
\examples{
move.from.preds.stepwise()

}
\keyword{movement}
