% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truth.R
\name{plot.truth}
\alias{plot.truth}
\title{A rsf score function}
\usage{
plot.truth(formu = NA, pop_size = NA, base_ = NA, pred_choice = NA,
  n_preds = NA, n_preds_original = NA, gridsize = NA, predictors_ = NA,
  ind_sd = NA, n_values = NA, uc_type = NA, mode = NA,
  param_list = list(NA), silent = T)
}
\arguments{
\item{formu}{formula without an response such as ~ x1 + x2}

\item{pop_size}{integer defining the number of individuals}

\item{base_}{vector of numerics as the parameters coefficients}

\item{pred_choice}{vector  of integers defining the set of predictors originally selected}

\item{n_preds}{integer defining total number of predictors, quadratic conditions and interactions}

\item{n_preds_original}{integer defining the number of predictors}

\item{gridsize}{integer determining size of the map}

\item{predictors_}{matrix of predictors}

\item{ind_sd}{integer as the standard deviation of the normal distribution with a mean of 1 that is multiplied with the parameter coefficients}

\item{n_values}{integer as the length of the min - max - sequence of a predictor (x-axis of an effect plot)}

\item{uc_type}{character defining the type of uncertainty, "both", "marg" and "cond" are possible}

\item{mode}{character defining wether effect plots should be generated or if the function is used inside a bootstrap. "plot" and "bootstrap" are possible}

\item{param_list}{list of parameters for shorter lines, see code for details}
}
\value{
In plot mode it returns a data.frame with quantiles suitable for ggplot2 and the plot itself, in  bootstrap mode it returns an array with non-aggregated rsf-scores.
}
\description{
This function generates rsf scores/effect plots
}
\examples{
plot.truth()

}
\keyword{rsf}
