% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(formu = f.species, df = track, model = "glm",
  n_bootstrap = 100, pop_size = pop.size, stratified = F)
}
\arguments{
\item{formu}{formula to be used in the model}

\item{df}{data frame to be used in the model}

\item{model}{type of the model, "glm" and "glmm" currently available, "bayes" coming soon}

\item{n_bootstrap}{number of bootstraps}

\item{stratified}{logical wether a stratified or a pooled bootstrap should be performed}

\item{pop.size}{number of individuals}
}
\value{
returns betas of the bootstrapped models
}
\description{
This function generates bootstrapped betas.
}
\examples{
bootstrap()

}
\keyword{bootstrap}
