% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(f_species = NA, df = NA, analysis = "glm", n_bootstrap = NA,
  pop_size = NA, stratified = F)
}
\arguments{
\item{f_species}{formula to be used in the analysis}

\item{df}{data frame to be used in the analysis}

\item{analysis}{type of the analysis, "glm" and "glmm" currently available, "bayes" coming soon}

\item{n_bootstrap}{number of bootstraps}

\item{pop_size}{number of individuals}

\item{stratified}{logical wether a stratified or a pooled bootstrap should be performed}
}
\value{
returns betas of the bootstrapped analysis
}
\description{
This function generates bootstrapped models with glm/glmm .
}
\examples{
bootstrap()

}
\keyword{bootstrap}
